requirements.txt
transformers
datasets
torch
scikit-learn
accelerate
wandb
ipywidgets
evaluate
python-dotenv